+++ webarchive/__init__.py
@@ -24,7 +24,7 @@ from .webarchive import WebArchive
 from .webresource import WebResource
 from .exceptions import WebArchiveError
 
-__version__ = "0.3.2"
+__version__ = "0.3.3"
 
 __all__ = ["__version__",
            "WebArchive", "WebResource", "WebArchiveError"]