+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Changed
+* Do not force a newline after the doctype in `HTMLRewriter.handle_decl()`.
 * Moved `test_extracted_archive_display` from the unit tests to a separate script.
 * Removed `test_extracted_archive_display`'s dependency on Tkinter.