+++ CHANGELOG.md
@@ -4,9 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.3.3] - 2021-11-05
 ### Added
 * Unit tests for HTML- and CSS-rewriting logic.
+* Build script for the Windows version of Webarchive Extractor.
 ### Changed
 * Clean up the `WebResource` class's internal API.
 * Do not force a newline after the doctype in `HTMLRewriter.handle_decl()`.
@@ -103,7 +104,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * Initial public release.
 
-[Unreleased]: https://github.com/bmjcode/pywebarchive/compare/v0.3.2...HEAD
+[Unreleased]: https://github.com/bmjcode/pywebarchive/compare/v0.3.3...HEAD
+[0.3.3]: https://github.com/bmjcode/pywebarchive/compare/v0.3.2...v0.3.3
 [0.3.2]: https://github.com/bmjcode/pywebarchive/compare/v0.3.1...v0.3.2
 [0.3.1]: https://github.com/bmjcode/pywebarchive/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/bmjcode/pywebarchive/compare/v0.2.4...v0.3.0